+++ src/SFA.DAS.FAAV2.UITests/GlobalUsing.cs
@@ -1,16 +1,16 @@
-global using SFA.DAS.FAAV2.UITests.Project.Helpers;
-global using TechTalk.SpecFlow;
-global using SFA.DAS.FrameworkHelpers;
-global using SFA.DAS.RAA.DataGenerator;
-global using SFA.DAS.UI.FrameworkHelpers;
-global using System.Linq;
+global using OpenQA.Selenium;
 global using SFA.DAS.ConfigurationBuilder;
+global using SFA.DAS.FAAV2.UITests.Project.Helpers;
+global using SFA.DAS.FAAV2.UITests.Project.Pages.ApplicationOverview;
+global using SFA.DAS.FrameworkHelpers;
 global using SFA.DAS.Login.Service;
 global using SFA.DAS.Login.Service.Project.Helpers;
-global using OpenQA.Selenium;
 global using SFA.DAS.Login.Service.Project.Tests.Pages;
-global using SFA.DAS.UI.Framework.TestSupport;
-global using SFA.DAS.FAAV2.UITests.Project.Pages;
+global using SFA.DAS.RAA.DataGenerator;
 global using SFA.DAS.RAA.DataGenerator.Project;
 global using SFA.DAS.UI.Framework;
+global using SFA.DAS.UI.Framework.TestSupport;
+global using SFA.DAS.UI.FrameworkHelpers;
 global using System;
+global using System.Linq;
+global using TechTalk.SpecFlow;
