+++ src/SFA.DAS.FAAV2.UITests/Project/Helpers/FAAStepsHelper.cs
@@ -4,14 +4,9 @@ namespace SFA.DAS.FAAV2.UITests.Project.Helpers;
 
 public class FAAStepsHelper(ScenarioContext context)
 {
-    private readonly TabHelper _tabHelper = context.Get<TabHelper>();
-    private readonly FAADataHelper _faaDataHelper = context.Get<FAADataHelper>();
-    private readonly ObjectContext _objectContext = context.Get<ObjectContext>();
-    private readonly string _faaBaseUrl = UrlConfig.FAAV2_AppSearch;
-
     public FAASignedInLandingBasePage GoToFAAHomePage()
     {
-        _tabHelper.GoToUrl(_faaBaseUrl);
+        context.Get<TabHelper>().GoToUrl(UrlConfig.FAAV2_AppSearch);
 
         if (new CheckFAASignedOutLandingPage(context).IsPageDisplayed())
         {
@@ -58,7 +53,6 @@ public class FAAStepsHelper(ScenarioContext context)
     }
 
     private FAA_ApplicationOverviewPage GoToFAAHomePageAndApply() => GoToFAAHomePage().SearchByReferenceNumber().Apply();
-
 
     public FAA_ApplicationOverviewPage ApplyForFirstVacancy(bool qualificationdetails, bool trainingCourse, bool job, bool workExperience, bool interviewSupport, bool disabilityConfident)
     {
@@ -166,157 +160,6 @@ public class FAAStepsHelper(ScenarioContext context)
                 .VerifyDisabilityConfidence_1();
         }
 
-
         return applicationFormPage;
     }
-
-    //    applicationFormPage.EnterEducation();
-    //    applicationFormPage.EnterStartedYear();
-    //    applicationFormPage.EnterFinishedYear();
-    //    applicationFormPage.EnterQualificationdetails(qualificationdetails);
-    //    applicationFormPage.EnterWorkExperience(workExperience);
-    //    applicationFormPage.EnterTrainingCourse(trainingCourse);
-    //    applicationFormPage.AnswerAdditionalQuestions();
-    //    applicationFormPage.EnterStrengths();
-    //    applicationFormPage.EnterSkills();
-    //    applicationFormPage.EnterHobbiesAndInterests();
-    //    applicationFormPage.ClickSaveAndContinue();
-    //    applicationFormPage.SelectAcceptSubmit();
-    //    applicationFormPage.SubmitApprenticeshipApplication();
-    //}
-
-    //public string GetApplicationStatus()
-    //{
-    //    return OpenFAAHomePageinNewtab()
-    //        .FindAnApprenticeship()
-    //        .SearchByReferenceNumber()
-    //        .View()
-    //        .ApplicationStatus();
-    //}
-
-    //public FAA_ApprenticeSearchPage FindAnApprenticeship() => GoToFAAHomePage().FindAnApprenticeship();
-
-    //public FAA_CreateAnAccountPage StartFAAAccountCreation()
-    //{
-    //    _tabHelper.GoToUrl(_faaBaseUrl);
-
-    //    return new FAA_Indexpage(context)
-    //        .GoToSignInPage()
-    //        .ClickCreateAnAccountLink();
-    //}
-
-    //public static void CreateFAAAccount(FAA_CreateAnAccountPage accountCreationPage)
-    //{
-    //    accountCreationPage.CreateAccount()
-    //        .EnterActivationCode()
-    //        .ClickSaveAndContinue()
-    //        .VerifyPhoneNumberVerificationText()
-    //        .EnterVerificationCode()
-    //        .VerifySuccessfulVerificationText();
-    //}
-
-    //public void CreateFAAAccountWithNoActivation(FAA_CreateAnAccountPage accountCreationPage)
-    //{
-    //    var (username, password, _, _) = _objectContext.GetFAALogin();
-    //    accountCreationPage.CreateAccount()
-    //        .ClickSignOut()
-    //        .SubmitUnactivatedLoginDetails(username, password);
-    //}
-
-    //public void WithdrawVacancy()
-    //{
-    //    SearchByReferenceNumber()
-    //        .View()
-    //        .Withdraw()
-    //        .YesWithdraw();
-    //}
-
-    //public void DismissNotification(string status)
-    //{
-    //    switch (status)
-    //    {
-    //        case "Successful":
-    //            GoToFAAHomePage().DismissSuccessfulNotification();
-    //            break;
-
-    //        case "Unsccessful":
-    //            GoToFAAHomePage().DismissUnsuccessfulNotification().ReadFeedback().VerifyReadFeedbackText();
-    //            break;
-    //    }
-    //}
-
-    //public void ApplyForAVacancy(string qualificationdetails, string workExperience, string trainingCourse, bool isSearchByCategory = false)
-    //{
-    //    var applicationFormPage = isSearchByCategory ? SearchByCategory().Apply() : SearchByReferenceNumber().Apply();
-
-    //    if (_objectContext.IsApprenticeshipVacancyType())
-    //    {
-    //        applicationFormPage.EnterEducation();
-    //        applicationFormPage.EnterStartedYear();
-    //        applicationFormPage.EnterFinishedYear();
-    //        applicationFormPage.EnterQualificationdetails(qualificationdetails);
-    //        applicationFormPage.EnterWorkExperience(workExperience);
-    //        applicationFormPage.EnterTrainingCourse(trainingCourse);
-    //        applicationFormPage.AnswerAdditionalQuestions();
-    //        applicationFormPage.EnterStrengths();
-    //        applicationFormPage.EnterSkills();
-    //        applicationFormPage.EnterHobbiesAndInterests();
-    //        applicationFormPage.ClickSaveAndContinue();
-    //        applicationFormPage.SelectAcceptSubmit();
-    //        applicationFormPage.SubmitApprenticeshipApplication();
-    //    }
-    //    else
-    //    {
-    //        applicationFormPage.EnterWorkExperience(workExperience);
-    //        applicationFormPage.EnterTrainingCourse(trainingCourse);
-    //        applicationFormPage.EnterQualificationdetails(qualificationdetails);
-    //        applicationFormPage.AnswerAdditionalQuestions();
-    //        applicationFormPage.SubmitTraineeshipApplication();
-    //    }
-    //}
-
-    //public void ApplyForTraineeship()
-    //{
-    //    ApplyForAVacancy(string.Empty, string.Empty, string.Empty, false);
-    //}
-
-    //public void CheckNationWideVacancies()
-    //{
-    //    FAA_ApprenticeSearchResultsPage _faaApprenticeshipSearchResultsPage = new(context);
-    //    _faaApprenticeshipSearchResultsPage.CheckSortOrderAndDistance();
-    //}
-
-    //public void CreateDraftApplication() => SearchByReferenceNumber().Apply().ClickSave();
-
-    //public FAA_ApprenticeSummaryPage ConfirmDraftVacancyDeletion() => new FAA_MyApplicationsHomePage(context).ConfirmVacancyDeletion().ConfirmDraftVacancyDeletion();
-
-    //public void ChangePersonalSettings()
-    //{
-    //    var signInpage = GoToSettingsPage().ChangeTheEmailIdSettings().ChangeEmailAddress();
-
-    //    signInpage.ConfirmEmailAddressUpdate();
-
-    //    _objectContext.SetFAAUsername(_faaDataHelper.ChangedEmailId);
-
-    //    GoToSettingsPage().ChangePhoneNumberSettings().EnterVerificationCode().VerifySuccessfulVerificationText();
-    //}
-
-    //private FAA_ApprenticeSummaryPage SearchByReferenceNumber()
-    //{
-    //    if (_objectContext.IsApprenticeshipVacancyType())
-    //        return FindAnApprenticeship().SearchByReferenceNumber();
-    //    else
-    //        return FindATraineeship().SearchByReferenceNumber();
-    //}
-
-    //private FAA_MyApplicationsHomePage OpenFAAHomePageinNewtab()
-    //{
-    //    _tabHelper.OpenInNewTab(_faaBaseUrl);
-
-    //    return new FAA_FindAnApprenticeshipHomePage(context).MyApplications();
-    //}
-
-    //private FAA_SettingsPage GoToSettingsPage() => GoToFAAHomePage().GoToSettings();
-
-    //private FAA_ApprenticeSummaryPage SearchByCategory() => FindAnApprenticeship().BrowseVacancy().SelectBrowsedVacancy();
 }
