+++ src/SFA.DAS.FAAV2.UITests/Project/Tests/Features/FAA_TestDataPrep_ApplyVacancy.feature
@@ -1,13 +1,11 @@
 ﻿Feature: FAA_TestDataPrep_ApplyVacancy
 
-@faav2
 @faaapplytestdataprep
 Scenario: FAA_TestDataPrep_ApplyFirstVacancy
 	Then the candidate can login in to faav2
 	And the candidate can apply for their first vacancy
 
 
-@faav2
 @faaapplytestdataprep
 Scenario: FAA_TestDataPrep_ApplyForVacancy
 	Then the candidate can login in to faav2
