+++ src/SFA.DAS.FAAV2.UITests/Project/Tests/Pages/FAASearchApprenticeLandingPage.cs
@@ -1,0 +1,8 @@
+﻿namespace SFA.DAS.FAAV2.UITests.Project.Tests.Pages;
+
+public class FAASearchApprenticeLandingPage(ScenarioContext context) : FAASignedInLandingBasePage(context)
+{
+    protected override By PageHeader => By.CssSelector(".govuk-heading-xl");
+
+    protected override string PageTitle => "Search apprenticeships";
+}
