--- src/SFA.DAS.FAAV2.UITests/Project/Tests/Pages/FAA_ApplicationOverviewPage2.cs
@@ -1,9 +1,0 @@
-﻿using SFA.DAS.FAAV2.UITests.Project.Tests.Pages;
-
-namespace SFA.DAS.FAAV2.UITests.Project.Pages;
-
-public partial class FAA_ApplicationOverviewPage(ScenarioContext context) : FAABasePage(context)
-{
-    protected override string PageTitle => $"Apply for {vacancyTitleDataHelper.VacancyTitle}";
-
-}
