+++ src/SFA.DAS.FAAV2.UITests/Project/Tests/Pages/FAA_ApplicationsPage.cs
@@ -1,0 +1,33 @@
+﻿namespace SFA.DAS.FAAV2.UITests.Project.Tests.Pages;
+
+public class FAA_ApplicationsPage(ScenarioContext context) : FAABasePage(context)
+{
+    protected override string PageTitle => "Your applications";
+
+    private static By SuccessfulLink => By.CssSelector("a[href='/applications?tab=Successful']");
+
+    private static By UnsuccessfulLink => By.CssSelector("a[href='/applications?tab=Unsuccessful']");
+
+    private static By SearchResultItem => By.CssSelector(".das-search-results__list-item");
+
+    private static By ViewApplicationLink => By.CssSelector("a[href*='view']");
+
+    public FAA_SuccessfulApplicationPage OpenSuccessfulApplicationPage()
+    {
+        formCompletionHelper.Click(SuccessfulLink);
+
+        return new(context);
+    }
+
+    public FAA_UnSuccessfulApplicationPage OpenUnSuccessfulApplicationPage()
+    {
+        formCompletionHelper.Click(UnsuccessfulLink);
+
+        return new(context);
+    }
+
+    public void ViewApplication()
+    {
+        pageInteractionHelper.FindElements(SearchResultItem).Single(x => x.Text.ContainsCompareCaseInsensitive(vacancyTitleDataHelper.VacancyTitle)).FindElement(ViewApplicationLink).Click();
+    }
+}
