+++ src/SFA.DAS.FAAV2.UITests/Project/Tests/Pages/FAA_SubmittedApplicationPage.cs
@@ -1,0 +1,8 @@
+﻿namespace SFA.DAS.FAAV2.UITests.Project.Tests.Pages;
+
+public class FAA_SubmittedApplicationPage(ScenarioContext context) : FAA_ApplicationsPage(context)
+{
+    protected override By PageHeader => By.CssSelector(".govuk-heading-m");
+
+    protected override string PageTitle => "Submitted";
+}
