+++ src/SFA.DAS.FAAV2.UITests/Project/Tests/Pages/StubPages/StubYouHaveSignedInFAAPage.cs
@@ -1,0 +1,10 @@
+﻿namespace SFA.DAS.FAAV2.UITests.Project.Tests.Pages.StubPages;
+
+public class StubYouHaveSignedInFAAPage(ScenarioContext context, string username, string idOrUserRef, bool newUser) : StubYouHaveSignedInBasePage(context, username, idOrUserRef, newUser)
+{
+    protected override By MainContent => By.CssSelector("[id='estimate-start-transfer']");
+
+    protected override By ContinueButton => By.CssSelector("[id='estimate-start-transfer'] a.govuk-button");
+
+    public new void Continue() => base.Continue();
+}
