+++ src/SFA.DAS.RAA-V2-Employer.UITests/Project/Tests/Features/E2EJourney/RV2_E_E2E_01.feature
@@ -5,7 +5,6 @@
 @v2_e2e
 @v2e_e2e
 @regression
-@faav2migrationtests
 Scenario: RV2_E_E2E_01 - Create an advert with registered name, Approve, Apply and make Application Successful
 	Given the Employer creates an advert by using a registered name
 	And the Reviewer Approves the vacancy
